requirements.txt
ctransformers 
sentence-transformers #for embedding model 
pinecone-client
langchain_huggingface
langchain
flask
pypdf
openai
tiktoken
pinecone
langchain
pdf2image
pypdf
python-dotenv
datasets
ragas

#-e .  # search for setup.py file 


